Revert "apply order book changes for hybrid amm cda"
This reverts commit 2e497f339d36c05b5cfbf3292d5ea82465af1266.
